Fix gemspec executables
lita-tweet was overwriting the executables with all of the files inside of `/bin` (ie. guard, rake, rspec). This blog post <http://bundler.io/blog/2015/03/20/moving-bins-to-exe.html> goes into detail about exactly why that happens, and suggests this change as a recommended fix. I tested it, and it does in fact fix the issue.